Gestion de la longueur maximum d'un pointeur dans my_putaddress

+++ Includes/Auxiliary_Functions/print_pointer_address.h
@@ -2,6 +2,6 @@
 # define PRINT_POINTER_ADDRESS_H
 # include "../Handlers/handlers.h"
 
-size_t		my_putaddress(unsigned long long nb, t_flag flag);
+size_t		my_putaddress(uintmax_t nb, t_flag flag);
 
 #endif
